pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
request2 = "^0.2"
requests = "^2.28.1"
python-binance = "^1.0.16"
pandas = "^1.5.1"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
